requirements.txt
streamlit
pandas
yfinance
matplotlib
langchain
langgraph
langchain-google-genai
langchain-community
tavily-python
python-dotenv